requirements.txt
selenium
html-testRunner
PyYAML